Move fieldmapper into its own views file
